requirements.txt
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.3
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
threadpoolctl==3.1.0
pandas==1.4.4
tensorflow==2.10.0